--- client/src/pages/a/index.tsx
@@ -1,5 +1,0 @@
-const A = () => {
-  return <>A</>;
-};
-
-export default A;
